requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.31.0
azure-functions==1.21.3
azure-identity==1.19.0
azure-search==1.0.0b2
azure-search-documents==11.5.1
azure-storage-blob==12.24.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
cryptography==43.0.3
debugpy==1.8.7
decorator==5.1.1
distro==1.9.0
dynamics365crm-python==1.0.2
executing==2.1.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
isodate==0.7.2
jedi==0.19.1
jiter==0.6.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
multidict==6.1.0
nest-asyncio==1.6.0
nltk==3.9.1
oauthlib==3.2.2
openai==1.52.2
packaging==24.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
portalocker==2.10.1
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
yarl==1.15.3

